changed location of isPrivacyPolicy function to utils dir as requested
